+++ src/main/java/net/geforcemods/securitycraft/datagen/ItemModelGenerator.java
@@ -13,6 +13,7 @@ import net.geforcemods.securitycraft.blocks.mines.BrushableMineBlock;
 import net.geforcemods.securitycraft.blocks.mines.DeepslateMineBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.IReinforcedBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedButtonBlock;
+import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedChiseledBookShelfBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedPistonBaseBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedStainedGlassPaneBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedWallBlock;
@@ -63,7 +64,7 @@ public class ItemModelGenerator extends ItemModelProvider {
 					reinforcedStainedPane(block);
 				else if (block instanceof ReinforcedWallBlock wall)
 					reinforcedWallInventory(block, wall.getVanillaBlock());
-				else if (block instanceof ReinforcedButtonBlock || block instanceof ReinforcedPistonBaseBlock)
+				else if (block instanceof ReinforcedButtonBlock || block instanceof ReinforcedPistonBaseBlock || block instanceof ReinforcedChiseledBookShelfBlock)
 					reinforcedBlockInventory(block);
 				else if (block instanceof IReinforcedBlock)
 					simpleReinforcedParent(block);